Separate the generic code_template from the specific code_generator
