+++ dp_wizard/utils/code_template.py
@@ -3,8 +3,8 @@ from pathlib import Path
 
 
 class Template:
-    def __init__(self, path, root=__file__, template=None):
-        if path is not None:
+    def __init__(self, path, root=None, template=None):
+        if path is not None and root is not None:
             self._path = f"_{path}.py"
             template_path = Path(root).parent / "no-tests" / self._path
             self._template = template_path.read_text()
